+++ src/components/modules/escrow/InitializeEscrowForm.tsx
@@ -10,30 +10,18 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-import {
-  Command,
-  CommandEmpty,
-  CommandGroup,
-  CommandInput,
-  CommandItem,
-  CommandList,
-} from "@/components/ui/command";
-import { Check, ChevronsUpDown } from "lucide-react";
-import { cn } from "@/lib/utils";
 import { FaRegTrashCan } from "react-icons/fa6";
-import { statusOptions } from "@/constants/escrow/StatusOptions";
 import { useInitializeEscrowHook } from "@/components/modules/escrow/hooks/initialize-escrow.hook";
+import PopoverCategory from "./components/PopoverCategory";
 
 export function InitializeEscrowForm() {
-  const { form, onSubmit, handleAddMilestone, handleRemoveMilestone } =
-    useInitializeEscrowHook();
-
-  const milestones = form.watch("milestones");
+  const {
+    form,
+    onSubmit,
+    handleAddMilestone,
+    handleRemoveMilestone,
+    milestones,
+  } = useInitializeEscrowHook();
 
   return (
     <Form {...form}>
@@ -129,53 +117,7 @@ export function InitializeEscrowForm() {
                 }}
               />
 
-              <Popover>
-                <PopoverTrigger asChild>
-                  <Button
-                    variant="outline"
-                    role="combobox"
-                    className="w-[200px] justify-between"
-                  >
-                    {milestone.status
-                      ? statusOptions.find(
-                          (option) => option.value === milestone.status,
-                        )?.label
-                      : "Select Status"}
-                    <ChevronsUpDown className="ml-2 h-4 w-4 opacity-50" />
-                  </Button>
-                </PopoverTrigger>
-                <PopoverContent className="w-[200px] p-0">
-                  <Command>
-                    <CommandInput placeholder="Search status..." />
-                    <CommandList>
-                      <CommandEmpty>No status found.</CommandEmpty>
-                      <CommandGroup>
-                        {statusOptions.map((option) => (
-                          <CommandItem
-                            key={option.value}
-                            value={option.value}
-                            onSelect={() => {
-                              const updatedMilestones = [...milestones];
-                              updatedMilestones[index].status = option.value;
-                              form.setValue("milestones", updatedMilestones);
-                            }}
-                          >
-                            <Check
-                              className={cn(
-                                "mr-2 h-4 w-4",
-                                milestone.status === option.value
-                                  ? "opacity-100"
-                                  : "opacity-0",
-                              )}
-                            />
-                            {option.label}
-                          </CommandItem>
-                        ))}
-                      </CommandGroup>
-                    </CommandList>
-                  </Command>
-                </PopoverContent>
-              </Popover>
+              <PopoverCategory milestone={milestone} index={index} />
 
               <Button
                 onClick={() => handleRemoveMilestone(index)}
